refactor(GCS+gRPC): split GrpcClient class
The `storage::internal::GrpcClient` class was getting too big, and it
would have grown even larger as we implement the metadata operations
over gRPC.  This PR splits the helper functions to parse resources (e.g.
`ObjectMetadata` or `ObjectAccessControl`) and/or requests.
