+++ google/cloud/storage/bucket_access_control.h
@@ -82,7 +82,6 @@ class BucketAccessControl : private internal::AccessControlCommon {
   }
 
   friend struct internal::BucketAccessControlParser;
-  friend class internal::GrpcClient;
 };
 
 std::ostream& operator<<(std::ostream& os, BucketAccessControl const& rhs);
