+++ google/cloud/storage/google_cloud_cpp_storage_grpc.bzl
@@ -20,7 +20,10 @@ google_cloud_cpp_storage_grpc_hdrs = [
     "grpc_plugin.h",
     "internal/grpc_client.h",
     "internal/grpc_configure_client_context.h",
+    "internal/grpc_object_access_control_parser.h",
+    "internal/grpc_object_metadata_parser.h",
     "internal/grpc_object_read_source.h",
+    "internal/grpc_object_request_parser.h",
     "internal/grpc_resumable_upload_session.h",
     "internal/grpc_resumable_upload_session_url.h",
     "internal/hybrid_client.h",
@@ -32,7 +35,10 @@ google_cloud_cpp_storage_grpc_hdrs = [
 google_cloud_cpp_storage_grpc_srcs = [
     "grpc_plugin.cc",
     "internal/grpc_client.cc",
+    "internal/grpc_object_access_control_parser.cc",
+    "internal/grpc_object_metadata_parser.cc",
     "internal/grpc_object_read_source.cc",
+    "internal/grpc_object_request_parser.cc",
     "internal/grpc_resumable_upload_session.cc",
     "internal/grpc_resumable_upload_session_url.cc",
     "internal/hybrid_client.cc",
