+++ google/cloud/storage/internal/common_metadata.h
@@ -60,7 +60,7 @@ inline bool operator>=(Owner const& lhs, Owner const& rhs) {
 }
 
 namespace internal {
-class GrpcClient;
+struct GrpcObjectMetadataParser;
 template <typename Derived>
 struct CommonMetadataParser;
 
@@ -101,7 +101,7 @@ class CommonMetadata {
   std::chrono::system_clock::time_point updated() const { return updated_; }
 
  private:
-  friend class GrpcClient;
+  friend struct GrpcObjectMetadataParser;
   template <typename ParserDerived>
   friend struct CommonMetadataParser;
