+++ google/cloud/storage/internal/grpc_object_access_control_parser.h
@@ -1,0 +1,43 @@
+// Copyright 2021 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     https://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+#ifndef GOOGLE_CLOUD_CPP_GOOGLE_CLOUD_STORAGE_INTERNAL_GRPC_OBJECT_ACCESS_CONTROL_PARSER_H
+#define GOOGLE_CLOUD_CPP_GOOGLE_CLOUD_STORAGE_INTERNAL_GRPC_OBJECT_ACCESS_CONTROL_PARSER_H
+
+#include "google/cloud/storage/internal/raw_client.h"
+#include "google/cloud/storage/version.h"
+#include <google/storage/v2/storage.pb.h>
+
+namespace google {
+namespace cloud {
+namespace storage {
+GOOGLE_CLOUD_CPP_INLINE_NAMESPACE_BEGIN
+namespace internal {
+
+struct GrpcObjectAccessControlParser {
+  static google::storage::v2::ObjectAccessControl ToProto(
+      ObjectAccessControl const& acl);
+  static ObjectAccessControl FromProto(
+      google::storage::v2::ObjectAccessControl acl,
+      std::string const& bucket_name, std::string const& object_name,
+      std::uint64_t generation);
+};
+
+}  // namespace internal
+GOOGLE_CLOUD_CPP_INLINE_NAMESPACE_END
+}  // namespace storage
+}  // namespace cloud
+}  // namespace google
+
+#endif  // GOOGLE_CLOUD_CPP_GOOGLE_CLOUD_STORAGE_INTERNAL_GRPC_OBJECT_ACCESS_CONTROL_PARSER_H
