+++ google/cloud/storage/storage_client_grpc_unit_tests.bzl
@@ -19,11 +19,12 @@
 storage_client_grpc_unit_tests = [
     "internal/grpc_client_failures_test.cc",
     "internal/grpc_client_insert_object_media_test.cc",
-    "internal/grpc_client_object_request_test.cc",
     "internal/grpc_client_read_object_test.cc",
-    "internal/grpc_client_test.cc",
     "internal/grpc_configure_client_context_test.cc",
+    "internal/grpc_object_access_control_parser_test.cc",
+    "internal/grpc_object_metadata_parser_test.cc",
     "internal/grpc_object_read_source_test.cc",
+    "internal/grpc_object_request_parser_test.cc",
     "internal/grpc_resumable_upload_session_test.cc",
     "internal/grpc_resumable_upload_session_url_test.cc",
     "internal/storage_auth_test.cc",
